requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
cssselect==1.0.3
fake-useragent==0.1.11
idna==2.8
lxml==4.9.1
parse==1.12.0
Pillow>=6.2.2
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.0
requests-html==0.10.0
six==1.12.0
soupsieve==1.9.2
tqdm==4.32.2
urllib3==1.26.5
w3lib==1.20.0
websockets==9.1
PyExecJS==1.5.1
